+++ docs/src/sections/GridClickEventSection.js
@@ -7,7 +7,13 @@ import Samples from '../Samples';
 export default function GridClickEventSection() {
   return (
     <div className="bs-docs-section">
-      <h3><Anchor id="grid-click-event">鼠标点击事件</Anchor></h3>
+      <h3><Anchor id="grid-click-event">Mouse click events</Anchor></h3>
+      <p>
+        The event handlers below are all triggered by events in the bubbling phase.
+        This means <code>onCellClick</code> will be called before <code>onRowClick</code>,
+        because <code>onCellClick</code> is registered on <code>td</code> tag
+        and <code>onRowClick</code> is registered on <code>tr</code> tag.
+      </p>
       <ReactPlayground codeText={Samples.GridClickEvent} />
     </div>
   );